versioning/requirements.txt
fastapi
fastapi_versioning
pydantic
pymongo
requests
mongo-queue-service
schedule
